requirements.txt
requests==2.26.0
google-cloud-storage==1.42.0
pytest
